:truck: chore: rename datagen utils to jsonl_convert
