+++ src/datagen/pfs_risco_tivea.py
@@ -6,7 +6,7 @@ from datetime import date, datetime
 import random 
 import re
 
-from utils import jsonl_data
+from jsonl_convert import jsonl_data
 
 faker = Faker('pt_BR')
 id_serial = itertools.count(start=0)
